--- src/utils/dates.js
@@ -1,14 +1,0 @@
-export const getEarliestDate = dates => {
-  const [earliest] = dates
-    .slice()
-    .sort((a, b) => new Date(a).getTime() - new Date(b).getTime())
-  return earliest
-}
-
-export const getLatestDate = dates => {
-  const [latest] = dates
-    .slice()
-    .sort((a, b) => new Date(a).getTime() - new Date(b).getTime())
-    .reverse()
-  return latest
-}
